src/requirements.txt
pyserial
requests
influxdb